requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
tabulate==0.9.0
click==8.1.7
rich==13.5.2
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.0